Fix file browser issues
Originally the open image / airunner path actions in the toolbar would
open the system's file browser. In order to run on Docker and to more
easily support Windows we are using QFileDialog instead of xdg. This
opens a consistent qt window that allows the user to interact with the
file system even when running on Docker.